# Do some speedups for varint encoding