--- a/encoder.go
+++ b/encoder.go
@@ -85,12 +85,12 @@ func (this *BinaryEncoder) WriteBoolean(x bool) {
 
 // Writes an int value.
 func (this *BinaryEncoder) WriteInt(x int32) {
-	this.buffer.Write(this.encodeVarint(int64(x)))
+	this.buffer.Write(this.encodeVarint32(x))
 }
 
 // Writes a long value.
 func (this *BinaryEncoder) WriteLong(x int64) {
-	this.buffer.Write(this.encodeVarint(x))
+	this.buffer.Write(this.encodeVarint64(x))
 }
 
 // Writes a float value.
@@ -152,8 +152,25 @@ func (this *BinaryEncoder) writeItemCount(count int64) {
 	this.WriteLong(count)
 }
 
-func (this *BinaryEncoder) encodeVarint(x int64) []byte {
-	var buf = make([]byte, 10)
+func (this *BinaryEncoder) encodeVarint32(n int32) []byte {
+	var buf [5]byte
+	ux := uint32(n) << 1
+	if n < 0 {
+		ux = ^ux
+	}
+	i := 0
+	for ux >= 0x80 {
+		buf[i] = byte(ux) | 0x80
+		ux >>= 7
+		i++
+	}
+	buf[i] = byte(ux)
+
+	return buf[0 : i+1]
+}
+
+func (this *BinaryEncoder) encodeVarint64(x int64) []byte {
+	var buf [10]byte
 	ux := uint64(x) << 1
 	if x < 0 {
 		ux = ^ux